Deprecate KnnVectoryQuery in favour of KnnFloatVectorQuery
