+++ lucene/core/src/java/org/apache/lucene/search/KnnVectorQuery.java
@@ -16,14 +16,7 @@
  */
 package org.apache.lucene.search;
 
-import java.io.IOException;
-import java.util.Arrays;
 import org.apache.lucene.codecs.KnnVectorsReader;
-import org.apache.lucene.document.KnnVectorField;
-import org.apache.lucene.index.FieldInfo;
-import org.apache.lucene.index.LeafReaderContext;
-import org.apache.lucene.index.VectorEncoding;
-import org.apache.lucene.util.ArrayUtil;
 import org.apache.lucene.util.Bits;
 
 /**
@@ -38,81 +31,17 @@ import org.apache.lucene.util.Bits;
  *   <li>Otherwise run a kNN search subject to the filter
  *   <li>If the kNN search visits too many vectors without completing, stop and run an exact search
  * </ul>
+ *
+ * @deprecated use {@link KnnFloatVectorQuery} instead
  */
-public class KnnVectorQuery extends AbstractKnnVectorQuery {
+@Deprecated
+public class KnnVectorQuery extends KnnFloatVectorQuery {
 
-  private static final TopDocs NO_RESULTS = TopDocsCollector.EMPTY_TOPDOCS;
-
-  private final float[] target;
-
-  /**
-   * Find the <code>k</code> nearest documents to the target vector according to the vectors in the
-   * given field. <code>target</code> vector.
-   *
-   * @param field a field that has been indexed as a {@link KnnVectorField}.
-   * @param target the target of the search
-   * @param k the number of documents to find
-   * @throws IllegalArgumentException if <code>k</code> is less than 1
-   */
   public KnnVectorQuery(String field, float[] target, int k) {
-    this(field, target, k, null);
+    super(field, target, k);
   }
 
-  /**
-   * Find the <code>k</code> nearest documents to the target vector according to the vectors in the
-   * given field. <code>target</code> vector.
-   *
-   * @param field a field that has been indexed as a {@link KnnVectorField}.
-   * @param target the target of the search
-   * @param k the number of documents to find
-   * @param filter a filter applied before the vector search
-   * @throws IllegalArgumentException if <code>k</code> is less than 1
-   */
   public KnnVectorQuery(String field, float[] target, int k, Query filter) {
-    super(field, k, filter);
-    this.target = target;
-  }
-
-  @Override
-  protected TopDocs approximateSearch(LeafReaderContext context, Bits acceptDocs, int visitedLimit)
-      throws IOException {
-    TopDocs results =
-        context.reader().searchNearestVectors(field, target, k, acceptDocs, visitedLimit);
-    return results != null ? results : NO_RESULTS;
-  }
-
-  @Override
-  VectorScorer createVectorScorer(LeafReaderContext context, FieldInfo fi) throws IOException {
-    if (fi.getVectorEncoding() != VectorEncoding.FLOAT32) {
-      return null;
-    }
-    return VectorScorer.create(context, fi, target);
-  }
-
-  @Override
-  public String toString(String field) {
-    return getClass().getSimpleName() + ":" + this.field + "[" + target[0] + ",...][" + k + "]";
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) return true;
-    if (super.equals(o) == false) return false;
-    KnnVectorQuery that = (KnnVectorQuery) o;
-    return Arrays.equals(target, that.target);
-  }
-
-  @Override
-  public int hashCode() {
-    int result = super.hashCode();
-    result = 31 * result + Arrays.hashCode(target);
-    return result;
-  }
-
-  /**
-   * @return the target query vector of the search. Each vector element is a float.
-   */
-  public float[] getTargetCopy() {
-    return ArrayUtil.copyOfSubArray(target, 0, target.length);
+    super(field, target, k, filter);
   }
 }
