+++ lucene/core/src/test/org/apache/lucene/document/TestManyKnnDocs.java
@@ -23,7 +23,7 @@ import org.apache.lucene.index.IndexWriterConfig;
 import org.apache.lucene.index.TieredMergePolicy;
 import org.apache.lucene.index.VectorSimilarityFunction;
 import org.apache.lucene.search.IndexSearcher;
-import org.apache.lucene.search.KnnVectorQuery;
+import org.apache.lucene.search.KnnFloatVectorQuery;
 import org.apache.lucene.search.TopDocs;
 import org.apache.lucene.store.Directory;
 import org.apache.lucene.store.FSDirectory;
@@ -65,7 +65,7 @@ public class TestManyKnnDocs extends LuceneTestCase {
       iw.forceMerge(1);
       iw.commit();
       IndexSearcher searcher = new IndexSearcher(DirectoryReader.open(dir));
-      TopDocs docs = searcher.search(new KnnVectorQuery("field", new float[] {120}, 10), 5);
+      TopDocs docs = searcher.search(new KnnFloatVectorQuery("field", new float[] {120}, 10), 5);
       assertEquals(5, docs.scoreDocs.length);
     }
   }
